works for all prime knots on <=11 crossings
there were some bugs, mainly involving the 'corners' of each vertical/horizontal bar as well as a problem in the size of the boundaries of certain 2-cells.
CURRENT LIMITATION: if, in an arc presentation, the end of some horizontal bar is joined to the beginning of the next immediate horizontal bar, the 2-cells will not have correct boundaries. This is an easy fix, but for now it's functional.